Revert "notfoundfilter-added"
This reverts commit 064b37d212b20c63abaa09a43b7eaa4b32d378be.

+++ Services/Categories/CategoryService.cs
@@ -80,22 +80,21 @@ namespace App.Services.Categories
 
         public async Task<ServiceResult> UpdateAsync(int id, UpdateCategoryRequest request)
         {
-            //var category = await categoryRepository.GetByIdAsync(id);
+            var category = await categoryRepository.GetByIdAsync(id);
 
-            //if (category == null)
-            //{
-            //    return ServiceResult.Fail("Category not found", System.Net.HttpStatusCode.NotFound);
-            //}
+            if (category == null)
+            {
+                return ServiceResult.Fail("Category not found", System.Net.HttpStatusCode.NotFound);
+            }
 
-            var iscategoryNameExist = await categoryRepository.Where(x => x.Name == request.Name && x.Id != id).AnyAsync();
+            var iscategoryNameExist = await categoryRepository.Where(x => x.Name == request.Name && x.Id != category.Id).AnyAsync();
 
             if (iscategoryNameExist)
             {
                 return ServiceResult.Fail("Category is already exists", System.Net.HttpStatusCode.BadRequest);
             }
 
-            var category = mapper.Map<Category>(request);
-            category.Id = id;
+            category = mapper.Map(request, category);
 
             categoryRepository.Update(category);
             await unitOfWork.SavechangesAsync();
@@ -108,10 +107,10 @@ namespace App.Services.Categories
         {
             var category = await categoryRepository.GetByIdAsync(id);
 
-            //if (category == null)
-            //{
-            //    return ServiceResult.Fail("Category not found", System.Net.HttpStatusCode.NotFound);
-            //}
+            if (category == null)
+            {
+                return ServiceResult.Fail("Category not found", System.Net.HttpStatusCode.NotFound);
+            }
 
             categoryRepository.Delete(category);
             await unitOfWork.SavechangesAsync();
