+++ Services/Products/ProductService.cs
@@ -123,15 +123,14 @@ namespace App.Services.Products
             // i dont response update and delete requests
             // i didnt write updateproductresponse record
 
-            //var product = await productRepository.GetByIdAsync(id);
-            
-            // i m checkin with filter (Notfoundfilter) 
-            //if (product is null)
-            //{
-            //    return ServiceResult.Fail("Product not found", HttpStatusCode.NotFound);
-            //}
+            var product = await productRepository.GetByIdAsync(id);
 
-            var isProductNameExist = await productRepository.Where(x => x.Name == request.Name && x.Id != id).AnyAsync();
+            if (product is null)
+            {
+                return ServiceResult.Fail("Product not found", HttpStatusCode.NotFound);
+            }
+
+            var isProductNameExist = await productRepository.Where(x => x.Name == request.Name && x.Id != product.Id).AnyAsync();
 
             if (isProductNameExist)
             {
@@ -142,12 +141,9 @@ namespace App.Services.Products
             //product.Stock = request.Stock;
             //product.Name = request.Name;
 
-            //product= mapper.Map(request, product);
+            product= mapper.Map(request, product);
 
-            var product = mapper.Map<Product>(request);
-            product.Id = id;
-
-            productRepository.Update(product!);
+            productRepository.Update(product);
             await unitOfWork.SavechangesAsync();
 
             // response is 204 no content
@@ -159,11 +155,10 @@ namespace App.Services.Products
         {
             var product = await productRepository.GetByIdAsync(request.ProductId);
 
-            // i m checkin with filter (Notfoundfilter) 
-            //if (product is null)
-            //{
-            //    return ServiceResult.Fail("Product not found",HttpStatusCode.NotFound);
-            //}
+            if (product is null)
+            {
+                return ServiceResult.Fail("Product not found",HttpStatusCode.NotFound);
+            }
 
             product.Stock = request.Quantity;
 
@@ -177,12 +172,10 @@ namespace App.Services.Products
         {
             var product = await productRepository.GetByIdAsync(id);
 
-            // i m checkin with filter (Notfoundfilter) 
-
-            //if (product is null)
-            //{
-            //    return ServiceResult.Fail("Product not found ", HttpStatusCode.NotFound);
-            //}
+            if (product is null)
+            {
+                return ServiceResult.Fail("Product not found ", HttpStatusCode.NotFound);
+            }
 
             productRepository.Delete(product);
